DropDown navbar became sticky
