+++ src/components/Header/Home.jsx
@@ -1,12 +1,11 @@
-import { useState } from "react";
 import Banner from "../Banner/Banner";
 import TapsiService from "../tapsi-service/TapsiService";
 import Content from "../content/Content";
 
-function Home({isNavOpen}) {
+function Home() {
   return (
     <div>
-      <Banner isNavOpen={isNavOpen}/>
+      <Banner/>
       <TapsiService/>
       <Content/>
     </div>
